Updated export coreml file

+++ tools/export_coreml.py
@@ -20,21 +20,18 @@ def patched_ms_deform_attn_forward(self, query, reference_points, value, value_s
 
     value = self.value_proj(value)
     if value_mask is not None:
-        value_mask = value_mask.astype(value.dtype).unsqueeze(-1)
+        value_mask = value_mask.to(value.dtype).unsqueeze(-1)
         value *= value_mask
     embed_dim = value.shape[-1]
     value = value.view(bs, Len_v, self.num_heads, self.head_dim)
 
     merged_q = Len_q * self.num_heads
 
-    # sampling_offsets (bs, Len_q, self.num_heads, self.num_levels, self.num_points, 2) -> (bs, merged_q, self.num_levels, self.num_points, 2)
     sampling_offsets = self.sampling_offsets(query).view(bs, Len_q, self.num_heads, self.num_levels, self.num_points, 2).flatten(1, 2)
 
-    # attention_weights (bs, Len_q, self.num_heads, self.num_levels*self.num_points) -> softmax -> (bs, merged_q, self.num_levels, self.num_points)
     attention_weights = self.attention_weights(query).view(bs, Len_q, self.num_heads, self.num_levels * self.num_points).flatten(1, 2)
     attention_weights = F.softmax(attention_weights, -1).view(bs, merged_q, self.num_levels, self.num_points)
 
-    # reference_points (bs, Len_q, self.num_levels, ref_dim) -> (bs, merged_q, self.num_levels, ref_dim)
     ref_dim = reference_points.shape[-1]
     reference_points = reference_points.repeat_interleave(self.num_heads, 1)
 
@@ -58,12 +55,10 @@ def patched_ms_deform_attn_forward(self, query, reference_points, value, value_s
 
         sampling_locations_l = sampling_locations[:, :, lvl, :, :]  # (bs, merged_q, self.num_points, 2)
 
-        # Normalize to -1 to 1
         sampling_grid_l = 2 * sampling_locations_l - 1
         sampling_grid_l = sampling_grid_l.flip(-1).view(bs, merged_q, self.num_points, 2)  # (bs, merged_q, p, 2) with (y, x)
 
-        # To make grid_sample work with 1D output, use H_out=1, W_out=p
-        sampled = F.grid_sample(value_l.repeat_interleave(merged_q // self.num_heads, dim=0), sampling_grid_l.view(bs * merged_q, 1, self.num_points, 2),
+        sampled = F.grid_sample(value_l.repeat_interleave(merged_q // bs, dim=0), sampling_grid_l.view(bs * merged_q, 1, self.num_points, 2),
                                 mode='bilinear', padding_mode='zeros', align_corners=False).squeeze(2)  # (bs * merged_q, head_dim, p)
 
         attn_l = attention_weights[:, :, lvl, :].view(bs * merged_q, 1, self.num_points)  # (bs * merged_q, 1, p)
@@ -73,7 +68,6 @@ def patched_ms_deform_attn_forward(self, query, reference_points, value, value_s
 
         output += out_l
 
-    # Reshape back
     output = output.view(bs, merged_q, self.head_dim)
     output = output.view(bs, Len_q, self.num_heads, self.head_dim).permute(0, 1, 2, 3).contiguous().view(bs, Len_q, embed_dim)
